structure MVC ok

+++ src/fr/spacey/utils/Position.java
@@ -1,33 +1,31 @@
 package fr.spacey.utils;
 
-import java.util.Vector;
-
 /**
  * Position dans l'espace avec deux coordonnées x et y
  * @author Benoit
  */
 public class Position {
 	
-	private int x, y;
+	private double x, y;
 
-	public Position(int x, int y) {
+	public Position(double x, double y) {
 		this.x = x;
 		this.y = y;
 	}
 
-	public int getX() {
+	public double getX() {
 		return x;
 	}
 
-	public int getY() {
+	public double getY() {
 		return y;
 	}
 	
-	public void setX(int x) {
+	public void setX(double x) {
 		this.x = x;
 	}
 
-	public void setY(int y) {
+	public void setY(double y) {
 		this.y = y;
 	}
